refactor: Move error case e2e tests to `/test` directory
