--- cmd/monaco/integrationtest/v2/test-resources/configs-with-invalid-payload/project/auto-tag/auto-tag-config-api.json
@@ -1,25 +1,0 @@
-{
-  "name": "{{ .name }}",
-  "rules": [
-    {
-      "type": "APPLICATION",
-      "enabled": true,
-      "valueFormat": null,
-      "propagationTypes": [],
-      "conditions": [
-        {
-          "key": {
-            "attribute": "WEB_APPLICATION_NAME"
-          },
-          "comparisonInfo": {
-            "type": "STRING",
-            "operator": "CONTAINS",
-            "value": "TEST",
-            "negate": false,
-            "caseSensitive": true
-          }
-        }
-      ]
-    }
-  ]
-}
