+++ test/errorcases/testdata/configs-with-duplicate-ids/manifest.yaml
