+++ test/errorcases/testdata/configs-with-duplicate-ids/project/alerting-profile/another_config.yaml
