+++ test/errorcases/testdata/configs-with-duplicate-ids/project/alerting-profile/config.yaml
