+++ test/errorcases/testdata/configs-with-duplicate-ids/project/alerting-profile/profile.json
