+++ test/errorcases/testdata/configs-with-duplicate-ids/project/auto-tag/config.yaml
