+++ test/errorcases/testdata/configs-with-duplicate-ids/project2/auto-tag/config.yaml
