+++ test/errorcases/testdata/configs-with-invalid-payload/manifest.yaml
