+++ test/errorcases/testdata/configs-with-invalid-payload/project/auto-tag/auto-tag-setting.json
