+++ test/errorcases/testdata/configs-with-invalid-payload/project/auto-tag/config.yaml
