+++ test/errorcases/testdata/deprecated-configs/manifest.yaml
