+++ test/errorcases/testdata/deprecated-configs/project/auto-tag.json
