+++ test/errorcases/testdata/deprecated-settings-schemas/manifest.yaml
