+++ test/errorcases/testdata/deprecated-settings-schemas/project/config.yaml
