+++ test/errorcases/testdata/deprecated-settings-schemas/project/span-attr.json
