+++ test/errorcases/testdata/deprecated-settings-schemas/project/span-event.json
