+++ test/errorcases/testdata/invalid-manifests/manifest_missing_projects.yaml
