+++ test/errorcases/testdata/invalid-manifests/manifest_non_existent_project.yaml
