+++ test/errorcases/testdata/invalid-manifests/project/alerting-profile/config.yaml
