+++ test/errorcases/testdata/invalid-manifests/project/alerting-profile/profile.json
